добавил абстрактный класс, наследование
