+++ src/components/button.js
@@ -1,4 +1,4 @@
-import {createElement} from "../utils.js";
+import AbstractComponent from './abstractComponent.js';
 
 const createLoadMoreButtonTemplate = function () {
   return (
@@ -6,21 +6,10 @@ const createLoadMoreButtonTemplate = function () {
   );
 };
 
-class LoadMoreButton {
-  constructor() {
-    this._element = null;
-  }
+class LoadMoreButton  extends AbstractComponent {
 
   getTemplate() {
     return createLoadMoreButtonTemplate();
-  }
-
-  getElement() {
-    if (!this._element) {
-      const template = this.getTemplate();
-      this._element = createElement(template);
-    }
-    return this._element;
   }
 }
